pc: define the series for the package
Add the ability to set the series for a snap.  We no longer have the
ubuntu series in the repository name, as this does not make semantic
sense.  The is particularly acute when we do not yet have a name for
the next LTS and we are producing snaps against a previous development
release.

Signed-off-by: Andy Whitcroft <apw@canonical.com>